refactor: StationListWindow 폴더 구조 개선
[#822]
