+++ frontend/src/components/ui/Navigator/NavigationBar/Menu.tsx
@@ -14,7 +14,7 @@ import Text from '@common/Text';
 
 import PersonalMenu from '@ui/Navigator/NavigationBar/PersonalMenu';
 import ServerStationFilters from '@ui/ServerStationFilters';
-import StationListWindow from '@ui/StationListWindow/StationListWindow';
+import StationListWindow from '@ui/StationListWindow';
 import StationSearchWindow from '@ui/StationSearchWindow';
 import LogoIcon from '@ui/Svg/LogoIcon';
 import LoginModal from '@ui/modal/LoginModal/LoginModal';
