+++ frontend/src/components/ui/StationListWindow/components/EmptyStationsNotice.tsx
