+++ frontend/src/components/ui/StationListWindow/components/StationSummaryCard.tsx
