+++ frontend/src/components/ui/StationListWindow/components/StationSummaryCardSkeleton.tsx
