+++ frontend/src/components/ui/StationListWindow/index.ts
@@ -1,0 +1,3 @@
+import StationListWindow from './StationListWindow';
+
+export default StationListWindow;
